Caleb holt
You're welcome, I like when I can contribute back to a project that has helped me.

I looked over the latest commit, and you wrote in the

CreateLaunch4jXMLTask.groovy
+                if(configuration.bundledJrePath != null)
+                    bundledJrePath(configuration.bundledJrePath)

I gave the change the benefit of the doubt and tried it in the launch4j GUI program but it threw an error when I attempted to use the xml generated by this code

Before I had it written like this

CreateLaunch4jXMLTask.groovy
+                if(configuration.bundledJrePath != null)
+                    xml.path(configuration.bundledJrePath)

It differs where you wrote bundledJrePath instead of xml.path. My change may have broken a convention you were going for where each attribute matched exactly as the XML wrote it. I guess I didn't really think about that at the time and wrote what the GUI and the docs called the configuration rather than how it was written in the xml.

Also, it's written as  xml.path(configuration.bundledJrePath)  instead of path(configuration.bundledJrePath)  because path is a method that is on the DefaultTask object so it needed to be specified that I was calling path on xml instead.

Thank you again, this project has helped mine a great deal.